Configure proxy-sigv4 segment internal plugin to prepare OSS release  (#2)
* chore: initialize repo with default backstage application

Create a generic Backstage app with minimal configuration to
kickstart plugin development. Utilized `npx @backstage/create-app@latest`
to generate the app.

* chore: move proxy-sigv4 segment internal plugin to prepare OSS release

Relocates proxy-sigv4 segment internal plugin to this repository
to prepare it for open source release. Additionally, the plugin
has been installed in the generic Backstage app for local testing
and verification.

* ci: create github action workflow to verify build and tests

The workflow verifies the build and tests for the entire repo,
including the plugins and the generic Backstage app.

* build: temporarily remove Node.js 20 from supported engine

Even though Backstage 1.19.x claims to have added support to Node.js 20, there are still incompatible dependencies that are causing builds to fail on this version.

* chore: add description to proxy-sigv4 package.json

* docs: add code of conduct and contribution expectation docs

* chore: adjust license copyright header and add to proxy-sigv4 entrypoint file